Printing Estimated Parameters for Explanatory IRT Models

This generic function prints estimated parameters from an eirm object returned from the eirm function.

# S3 method for eirm
print(x, difficulty = FALSE, ...)

Arguments

x
An object returned from the eirm function.

difficulty
Whether difficulty should be used instead of easiness (default: FALSE)

...
Other print-related arguments.

Value

Estimated parameters from an eirm object.

Examples

data("VerbAgg")
mod0 <- eirm(formula = "r2 ~ -1 + situ + btype + (1|id)", data = VerbAgg)
print(mod0) # or, just mod0

#> EIRM formula: "r2 ~ -1 + situ + btype + (1|id)" 
#> 
#> Number of persons: 316 
#> 
#> Number of observations: 7584 
#> 
#> Number of predictors: 4 
#> 
#> Parameter Estimates:
#> 
#>            Easiness   S.E. z-value   p-value
#> situother     1.376 0.0941   14.62  2.06e-48
#> situself      0.371 0.0913    4.06  4.85e-05
#> btypescold   -1.031 0.0671  -15.35  3.32e-53
#> btypeshout   -1.996 0.0718  -27.80 4.19e-170
#> 
#> Note: The estimated parameters above represent 'easiness'.
#> Use difficulty = TRUE to get difficulty parameters.

print(mod0, difficulty = TRUE)

#> EIRM formula: "r2 ~ -1 + situ + btype + (1|id)" 
#> 
#> Number of persons: 316 
#> 
#> Number of observations: 7584 
#> 
#> Number of predictors: 4 
#> 
#> Parameter Estimates:
#> 
#>            Difficulty   S.E. z-value   p-value
#> situother      -1.376 0.0941   14.62  2.06e-48
#> situself       -0.371 0.0913    4.06  4.85e-05
#> btypescold      1.031 0.0671  -15.35  3.32e-53
#> btypeshout      1.996 0.0718  -27.80 4.19e-170
#> 
#> Note: The estimated parameters above represent 'difficulty'.